Sar_app/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
timm
huggingface_hub
Pillow
python-multipart
openai
